RALPH: Mega-mock decomposition: tasks (#252)
Internal split of the 1700-line `tasks-mock-data.ts` into focused
sub-modules behind an unchanged public surface. Following the precedent
set by the seed extractions in #246–#250 (workspace splits), the seed
data lives in `seeds/tasks.ts`; structural code lives in `tasks-mock/`.

Layout
- src/data/seeds/tasks.ts (1443 lines): SEED_TASKS roster (66 tasks
  spanning all four statuses), assignee/item references, SeedQuestion
  / SeedTask types, and `hydrateSeedTasks` (which threads
  ORMATEK_SUPPLIERS into supplierQuestions). Exports
  `SEED_TASKS_HYDRATED: readonly Task[]`, computed once at module
  load. Mirrors the `seeds/items.ts` / `seeds/companies.ts` /
  `seeds/emails.ts` / `seeds/workspace-employees.ts` pattern. The
  seed file exceeds the 400-line target — it's flat data with
  inline section comments, splitting it into per-status files would
  fragment one logical roster and add navigation friction without
  reducing reader load (precedent: the suppliers Ormatek seed in
  #253 is explicitly exempted).
- src/data/tasks-mock/store.ts (62 lines): mutable `tasksStore`
  array, `cloneTask` defensive deep-copy, `seedStore`,
  `_resetTasksStore` / `_setTasks` / `_getAllTasks` public helpers,
  and the read/write accessors `readTasks` / `writeTaskAt` /
  `getTaskAt` plus `findTaskIndex` / `requireTaskIdx`. Mutations
  call `writeTaskAt(idx, task)` instead of indexing the live array
  directly so the store remains the only file that touches its
  internal mutable state.
- src/data/tasks-mock/queries.ts (116 lines): `fetchTaskBoardMock`,
  `fetchAllTasksMock`, `fetchTasksMock`, `fetchTaskMock`. Owns the
  `applyFilters` / `sortTasks` / `applySortIfAny` / `buildColumn`
  helpers and the `COLUMN_PAGE_SIZE` / `LIST_PAGE_SIZE` constants.
  Imports board/list response types from `domains/tasks` (canonical
  home) — the duplicate type declarations that used to live in
  tasks-mock-data are gone.
- src/data/tasks-mock/mutations.ts (60 lines): `changeTaskStatusMock`,
  `uploadTaskAttachmentsMock`, `deleteTaskAttachmentMock`. Owns the
  `fileExtension` helper. Reaches into the store via
  `requireTaskIdx` + `getTaskAt` + `writeTaskAt`.
- src/data/tasks-mock-data.ts (19 lines): now a barrel re-exporting
  the same 10 functions (`_getAllTasks`, `_resetTasksStore`,
  `_setTasks`, `fetchAllTasksMock`, `fetchTaskBoardMock`,
  `fetchTaskMock`, `fetchTasksMock`, `changeTaskStatusMock`,
  `deleteTaskAttachmentMock`, `uploadTaskAttachmentsMock`) that the
  in-memory adapter and the legacy mock-store test rely on.

Decisions
- Singleton not closure-isolated. The store is still a module-level
  singleton wrapped by `createInMemoryTasksClient`. Closure isolation
  doesn't unlock anything new for the tasks domain (no cross-store
  reach-ins from anywhere else in the codebase). Migrating the
  in-memory adapter to closure isolation is a follow-up that's
  natural to do whenever the next consumer needs an isolated store.
- `cloneTask` lives in `store.ts`, not in a shared util. It's the
  store's defensive-copy invariant — colocating it with the store
  keeps the contract obvious.
- Read accessor returns the live array reference (`readTasks(): Task[]`)
  rather than a copy. Queries map → cloneTask only when constructing
  the response payload, matching the original behavior. The doc on
  `readTasks` calls out the no-mutate contract.
- Duplicate response types (`BoardColumn`, `TaskBoardResponse`,
  `FetchTaskBoardParams`, `FetchTasksParams`, `TaskListResponse`)
  that used to be re-declared in tasks-mock-data are now imported
  from `domains/tasks` (where they were already canonical). Nothing
  imported them from tasks-mock-data, so this is a quiet cleanup —
  no consumer-facing impact.

Public surface unchanged
- `TasksClient` interface: untouched.
- `createInMemoryTasksClient` import surface: untouched (still
  `from "../tasks-mock-data"`).
- All 10 mock functions exported by `tasks-mock-data.ts` resolve to
  the same implementations.
- Test file `src/data/tasks-mock-data.test.ts` runs unmodified —
  still imports from `./tasks-mock-data`, still gets the same 10
  functions.

CONTEXT.md
- In-memory adapter blurb extended with the `tasks-mock/` layout
  (`store.ts` / `queries.ts` / `mutations.ts`) and the `seeds/tasks.ts`
  pointer.

Verified
- pnpm typecheck / lint / test / knip / build all clean.
- 1432 tests passing (unchanged from baseline — pure refactor, no
  test changes).

Acceptance criteria per #252
- Tasks in-memory adapter decomposed into store / queries / mutations
  / seed sub-modules ✓
- No file in the new structure exceeds a reasonable size (target:
  under 400 lines per module) ✓ for the structural modules
  (store 62 / queries 116 / mutations 60 / barrel 19); the seed
  (1443 lines) is flat data and intentionally one file per the
  rationale above
- Public client surface unchanged ✓
- All existing tests pass without modification ✓
- pnpm typecheck / pnpm lint / pnpm knip pass ✓

Notes for next iteration
- Closure isolation for the tasks in-memory adapter is the natural
  next refactor (mirror what #250 did for profile / company-info /
  workspace-employees / invitations). It needs to come with a small
  consumer-side migration: tests currently call `_setTasks` /
  `_resetTasksStore` directly; closure isolation moves seeding to
  the factory's `seed` option (already supported on
  `createInMemoryTasksClient`).
- The `_setTasks` / `_resetTasksStore` underscore reach-ins remain
  in `tasks-mock-data.test.ts` because it tests the underlying mock
  store, not the seam — same exception applied to items / suppliers
  / folders / notifications mock-module tests.
- Items / suppliers mock files are the remaining mega-mocks. The
  suppliers split (#253) is the obvious next slice — suppliers is
  larger than tasks (~3000 lines structural + 7000 lines Ormatek
  seed), and the Ormatek seed isolation is explicitly part of that
  AC.

Closes #252.
