+++ src/data/tasks-mock-data.ts
@@ -1,1704 +1,19 @@
-import { createBlobUrl, delay, nextId, paginate } from "./mock-utils";
-import { ORMATEK_SUPPLIERS } from "./suppliers-ormatek";
-import type { Attachment, Task, TaskFilterParams, TaskStatus } from "./task-types";
+/**
+ * Tasks mock data — barrel re-exporting the decomposed sub-modules. The
+ * actual store, queries, mutations, and seed live in `tasks-mock/` and
+ * `seeds/tasks.ts` respectively. This file preserves the import surface
+ * the in-memory adapter and the legacy mock-store test rely on.
+ */
 
-// --- Seed data ---
-
-const assigneeIvan = {
-	id: "user-1",
-	firstName: "Иван",
-	lastName: "Журавлёв",
-	email: "ivan.zhuravlyov.58@mostholding.ru",
-	avatarIcon: "blue",
-};
-const assigneeOlga = {
-	id: "user-2",
-	firstName: "Ольга",
-	lastName: "Соколова",
-	email: "o.sokolova@ormatek.com",
-	avatarIcon: "green",
-};
-const assigneeDmitry = {
-	id: "user-3",
-	firstName: "Дмитрий",
-	lastName: "Орлов",
-	email: "d.orlov@ormatek.com",
-	avatarIcon: "orange",
-};
-const assigneeEkaterina = {
-	id: "user-4",
-	firstName: "Екатерина",
-	lastName: "Белова",
-	email: "e.belova@ormatek.com",
-	avatarIcon: "red",
-};
-
-const ITEM_PVD = { id: "item-1", name: "Полотно ПВД 2600 мм", companyId: "company-1" };
-const ITEM_PPU = { id: "item-2", name: "Пенополиуретан ППУ ST2240", companyId: "company-1" };
-const ITEM_FABRIC = { id: "item-3", name: "Ткань жаккардовая 360 г/м²", companyId: "company-1" };
-const ITEM_LDSP = { id: "item-4", name: "ЛДСП 16 мм, дуб сонома", companyId: "company-1" };
-
-type SeedQuestion = { id: string; question: string; answer: string | null };
-type SeedTask = Omit<Task, "supplierQuestions"> & { supplierQuestions: SeedQuestion[] };
-
-const SEED_TASKS: SeedTask[] = [
-	// --- Assigned (6) ---
-	{
-		id: "task-1",
-		name: "Список документов для договора",
-		status: "assigned",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-08T15:42:00.000Z",
-		deadlineAt: "2026-04-22T18:00:00.000Z",
-		description: "Поставщики просят уточнить перечень документов, необходимых для заключения договора поставки.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-1-1", question: "Какие документы нужны со стороны ОРМАТЕК для заключения договора?", answer: null },
-			{ id: "sq-1-2", question: "Требуется ли доверенность на получение товара?", answer: null },
-			{ id: "sq-1-3", question: "Нужна ли карточка предприятия?", answer: null },
-		],
-		updatedAt: "2026-04-08T15:42:00.000Z",
-	},
-	{
-		id: "task-2",
-		name: "Срок принятия решения по КП",
-		status: "assigned",
-		item: ITEM_PPU,
-		assignee: assigneeOlga,
-		createdAt: "2026-02-27T11:48:00.000Z",
-		deadlineAt: "2026-04-20T18:00:00.000Z",
-		description:
-			"Несколько поставщиков уточняют, в какие сроки ОРМАТЕК готов принять окончательное решение по предложениям.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-2-1", question: "До какого числа действует срок рассмотрения КП?", answer: null },
-			{ id: "sq-2-2", question: "Будет ли возможность скорректировать цену после первичной оценки?", answer: null },
-			{ id: "sq-2-3", question: "Какие критерии выбора поставщика являются ключевыми?", answer: null },
-		],
-		updatedAt: "2026-02-27T11:48:00.000Z",
-	},
-	{
-		id: "task-3",
-		name: "Приглашение к ЭДО",
-		status: "assigned",
-		item: ITEM_FABRIC,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-13T16:48:00.000Z",
-		deadlineAt: "2026-04-25T18:00:00.000Z",
-		description: "Поставщик просит прислать приглашение к ЭДО в Диадоке для обмена закрывающими документами.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [{ id: "sq-3-1", question: "В каком операторе ЭДО работает ОРМАТЕК?", answer: null }],
-		updatedAt: "2026-04-13T16:48:00.000Z",
-	},
-	{
-		id: "task-4",
-		name: "Отсутствует текст запроса в письме (ООО Профессиональные Решения)",
-		status: "assigned",
-		item: ITEM_LDSP,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-08T12:29:00.000Z",
-		deadlineAt: "2026-04-18T18:00:00.000Z",
-		description: "Поставщик получил письмо без содержательной части запроса. Нужно повторно отправить спецификацию.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [{ id: "sq-4-1", question: "Можно ли повторно прислать ТЗ с характеристиками?", answer: null }],
-		updatedAt: "2026-04-08T12:29:00.000Z",
-	},
-	{
-		id: "task-5",
-		name: "Цвет плёнки (прозрачность)",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-25T15:17:00.000Z",
-		deadlineAt: "2026-04-20T18:00:00.000Z",
-		description: "Поставщики уточняют цвет плёнки — прозрачная или матовая, допустим ли лёгкий молочный оттенок.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-5-1", question: "Допустим ли молочный оттенок при условии сохранения первичного сырья?", answer: null },
-		],
-		updatedAt: "2026-02-25T15:17:00.000Z",
-	},
-	{
-		id: "task-6",
-		name: "Уточнение параметров плёнки и объёмов (КРЕДО)",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-10T11:05:00.000Z",
-		deadlineAt: "2026-04-24T18:00:00.000Z",
-		description: "КРЕДО просит подтвердить плотность, толщину и помесячный график отгрузок.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-6-1", question: "Какая требуется толщина плёнки (мкм)?", answer: null },
-			{ id: "sq-6-2", question: "Каков помесячный график отгрузок в течение года?", answer: null },
-		],
-		updatedAt: "2026-04-10T11:05:00.000Z",
-	},
-	// --- Completed (10) ---
-	{
-		id: "task-7",
-		name: "Уточнение объёма одной партии для расчёта доставки",
-		status: "completed",
-		item: ITEM_PPU,
-		assignee: assigneeDmitry,
-		createdAt: "2026-02-18T11:59:00.000Z",
-		deadlineAt: "2026-02-25T18:00:00.000Z",
-		description: "Поставщик запрашивает объём одной поставки для расчёта стоимости транспортной логистики.",
-		questionCount: 3,
-		completedResponse:
-			"Партия — 15 000 м (≈1 500 кг), ежемесячная отгрузка. Доставка — до производства в Аксайском р-не Ростовской обл.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-7-1", question: "Объём одной партии?", answer: "15 000 м" },
-			{ id: "sq-7-2", question: "Периодичность поставок?", answer: "Ежемесячно" },
-			{ id: "sq-7-3", question: "Адрес доставки?", answer: "Ростовская обл., Аксайский р-н, Южная промзона" },
-		],
-		updatedAt: "2026-02-20T14:30:00.000Z",
-	},
-	{
-		id: "task-8",
-		name: "Согласование альтернативы: полурукав 1,3 м вместо полотна 2,6 м",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-19T12:54:00.000Z",
-		deadlineAt: "2026-02-26T18:00:00.000Z",
-		description: "Часть поставщиков предлагает полурукав 1 300 мм (с разворотом до 2 600 мм) как аналог полотна.",
-		questionCount: 2,
-		completedResponse:
-			"Полурукав с разворотом до 2 600 мм допустим при условии первичного сырья и соответствующей толщины. Цену сравниваем с полотном по цене за 1 кг.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-8-1", question: "Допустим ли полурукав 1,3 м с разворотом?", answer: "Да, при первичном сырье" },
-			{ id: "sq-8-2", question: "Как сравниваете цену?", answer: "За 1 кг материала" },
-		],
-		updatedAt: "2026-02-21T10:15:00.000Z",
-	},
-	{
-		id: "task-9",
-		name: "Рассмотрение предложения без НДС (Polystan)",
-		status: "completed",
-		item: ITEM_FABRIC,
-		assignee: assigneeOlga,
-		createdAt: "2026-02-20T12:43:00.000Z",
-		deadlineAt: "2026-02-27T18:00:00.000Z",
-		description: "Polystan работает по УСН без НДС — уточнить, принимаем ли предложения от таких поставщиков.",
-		questionCount: 1,
-		completedResponse: "ОРМАТЕК принимает предложения и без НДС. Сравнение КП — по итоговой цене с учётом доставки.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-9-1", question: "Рассматриваем ли предложения без НДС?", answer: "Да, по итоговой цене с доставкой" },
-		],
-		updatedAt: "2026-02-22T09:40:00.000Z",
-	},
-	{
-		id: "task-10",
-		name: "Уточнение условий договора и объёмов для Армодверь",
-		status: "completed",
-		item: ITEM_LDSP,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-02-18T12:03:00.000Z",
-		deadlineAt: "2026-02-25T18:00:00.000Z",
-		description: "Поставщик Армодверь уточняет годовой объём и условия оплаты.",
-		questionCount: 1,
-		completedResponse: "Годовой объём — 180 000 м. Оплата — 100% предоплата в течение 2 банковских дней.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-10-1", question: "Годовой объём и условия оплаты?", answer: "180 000 м, 100% предоплата (2 б. д.)" },
-		],
-		updatedAt: "2026-02-19T11:00:00.000Z",
-	},
-	{
-		id: "task-11",
-		name: "Уточнение характеристик плёнки и доставки для ООО «Арли»",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeDmitry,
-		createdAt: "2026-02-16T09:38:00.000Z",
-		deadlineAt: "2026-02-23T18:00:00.000Z",
-		description: "ООО «Арли» просит полное ТЗ: ширина, толщина, цвет и условия доставки.",
-		questionCount: 2,
-		completedResponse:
-			"ПВД 2 600 мм, прозрачная, первичка. Доставка до Аксайского р-на включается в счёт. Толщина — по согласованию.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-11-1", question: "Точные характеристики плёнки?", answer: "ПВД 2 600 мм, первичка, прозрачная" },
-			{ id: "sq-11-2", question: "Условия доставки?", answer: "До производства в Аксайском р-не, включена в счёт" },
-		],
-		updatedAt: "2026-02-17T16:20:00.000Z",
-	},
-	{
-		id: "task-12",
-		name: "Уточнение точного адреса доставки (номер дома/склада)",
-		status: "completed",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-17T10:12:00.000Z",
-		deadlineAt: "2026-02-24T18:00:00.000Z",
-		description: "Транспортные компании поставщиков запрашивают точные координаты разгрузки.",
-		questionCount: 1,
-		completedResponse:
-			"Ростовская обл., Аксайский р-н, Грушевское с/п, Южная промзона, склад №3. Контакт: Королёв С., +7 (863) 320-01-01.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-12-1", question: "Точный адрес и контакт на складе?", answer: "Южная промзона, склад №3, Королёв С." },
-		],
-		updatedAt: "2026-02-18T13:40:00.000Z",
-	},
-	{
-		id: "task-13",
-		name: "Подтверждение первичного сырья (без вторсырья)",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeOlga,
-		createdAt: "2026-02-15T14:20:00.000Z",
-		deadlineAt: "2026-02-22T18:00:00.000Z",
-		description: "Часть поставщиков предлагает ТУ-сырьё. Нужно подтвердить требование к первичке.",
-		questionCount: 1,
-		completedResponse: "Требование — только первичное сырьё (без вторсырья). Предложения на ТУ-сырьё не рассматриваем.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [{ id: "sq-13-1", question: "Допустимо ли ТУ-сырьё?", answer: "Нет, только первичка" }],
-		updatedAt: "2026-02-16T12:00:00.000Z",
-	},
-	{
-		id: "task-14",
-		name: "Запрос образца плёнки перед заключением договора",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-02-20T10:50:00.000Z",
-		deadlineAt: "2026-02-28T18:00:00.000Z",
-		description: "Нужен ли образец плёнки для ОТК перед первой отгрузкой.",
-		questionCount: 1,
-		completedResponse: "Образец 2×2 м запрашиваем у финалистов — после отбора по цене и условиям оплаты.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-14-1", question: "Когда и какой образец нужен?", answer: "2×2 м после отбора финалистов" },
-		],
-		updatedAt: "2026-02-22T15:10:00.000Z",
-	},
-	{
-		id: "task-15",
-		name: "Согласование условий отсрочки оплаты",
-		status: "completed",
-		item: ITEM_FABRIC,
-		assignee: assigneeDmitry,
-		createdAt: "2026-02-22T11:30:00.000Z",
-		deadlineAt: "2026-03-01T18:00:00.000Z",
-		description: "Поставщики уточняют, рассматривает ли ОРМАТЕК отсрочку 14–30 дней вместо предоплаты.",
-		questionCount: 2,
-		completedResponse:
-			"Первая поставка — 100% предоплата. После трёх успешных отгрузок — обсуждается отсрочка до 14 дней.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-15-1", question: "Возможна ли отсрочка?", answer: "Со второго контракта — до 14 дней" },
-			{ id: "sq-15-2", question: "Когда пересматриваются условия?", answer: "После 3 успешных отгрузок" },
-		],
-		updatedAt: "2026-02-24T09:20:00.000Z",
-	},
-	{
-		id: "task-16",
-		name: "Сравнение цены за 1 кг vs за 1 м у разных поставщиков",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-24T16:00:00.000Z",
-		deadlineAt: "2026-03-03T18:00:00.000Z",
-		description: "Поставщики указывают цены по-разному: кто-то за 1 кг, кто-то за 1 м. Нужен единый подход.",
-		questionCount: 1,
-		completedResponse:
-			"Для итогового сравнения приводим все предложения к цене за 1 кг (с учётом толщины и плотности).",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-16-1", question: "Как приводим цены к общему знаменателю?", answer: "К цене за 1 кг" },
-		],
-		updatedAt: "2026-02-26T10:40:00.000Z",
-	},
-	// --- Assigned (12 more) ---
-	{
-		id: "task-17",
-		name: "Условия хранения плёнки на складе поставщика",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-15T09:30:00.000Z",
-		deadlineAt: "2026-04-29T18:00:00.000Z",
-		description: "Нужно уточнить условия хранения рулонов до отгрузки: температура, влажность, защита от УФ.",
-		questionCount: 4,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-17-1", question: "При какой температуре хранятся рулоны?", answer: null },
-			{ id: "sq-17-2", question: "Какая максимальная влажность допустима?", answer: null },
-			{ id: "sq-17-3", question: "Есть ли защита от прямого солнечного света?", answer: null },
-			{ id: "sq-17-4", question: "Максимальный срок хранения до отгрузки?", answer: null },
-		],
-		updatedAt: "2026-04-15T09:30:00.000Z",
-	},
-	{
-		id: "task-18",
-		name: "Полный пакет сертификатов (ООО «ПолиТекс»)",
-		status: "assigned",
-		item: ITEM_PPU,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-14T13:12:00.000Z",
-		deadlineAt: "2026-04-28T18:00:00.000Z",
-		description:
-			"«ПолиТекс» запрашивает список сертификатов, которые должен предоставить поставщик для заключения договора.",
-		questionCount: 7,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-18-1", question: "Требуется ли сертификат соответствия ГОСТ?", answer: null },
-			{ id: "sq-18-2", question: "Нужен ли СЭЗ (санитарно-эпидемиологическое заключение)?", answer: null },
-			{ id: "sq-18-3", question: "Требуется ли паспорт качества на каждую партию?", answer: null },
-			{ id: "sq-18-4", question: "Нужен ли сертификат ISO 9001?", answer: null },
-			{ id: "sq-18-5", question: "Требуется ли декларация о соответствии ТР ТС?", answer: null },
-			{ id: "sq-18-6", question: "Нужен ли протокол испытаний по толщине?", answer: null },
-			{ id: "sq-18-7", question: "Принимается ли копия сертификата или только оригинал?", answer: null },
-		],
-		updatedAt: "2026-04-14T13:12:00.000Z",
-	},
-	{
-		id: "task-19",
-		name: "Маркировка рулонов (шевронная или стандартная)",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-16T10:00:00.000Z",
-		deadlineAt: "2026-04-30T18:00:00.000Z",
-		description: "Уточнить требования к маркировке рулонов: номер партии, дата, цвет этикетки.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-19-1", question: "Какая информация должна быть на этикетке?", answer: null },
-			{ id: "sq-19-2", question: "Требуется ли штрих-код на каждом рулоне?", answer: null },
-		],
-		updatedAt: "2026-04-16T10:00:00.000Z",
-	},
-	{
-		id: "task-20",
-		name: "Минимальная партия для расчёта цены",
-		status: "assigned",
-		item: ITEM_FABRIC,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-17T11:20:00.000Z",
-		deadlineAt: "2026-05-01T18:00:00.000Z",
-		description: "Поставщики спрашивают минимальный заказ для специальной цены.",
-		questionCount: 5,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-20-1", question: "Какой минимальный объём заказа?", answer: null },
-			{ id: "sq-20-2", question: "Есть ли скидка за 6-месячный контракт?", answer: null },
-			{ id: "sq-20-3", question: "Зависит ли цена от времени года?", answer: null },
-			{ id: "sq-20-4", question: "Возможна ли фиксация цены на квартал?", answer: null },
-			{ id: "sq-20-5", question: "При каком объёме начинается оптовая скидка?", answer: null },
-		],
-		updatedAt: "2026-04-17T11:20:00.000Z",
-	},
-	{
-		id: "task-21",
-		name: "Технические характеристики втулки рулона",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-18T14:50:00.000Z",
-		deadlineAt: "2026-05-02T18:00:00.000Z",
-		description: "Поставщики хотят согласовать размер и материал втулки для совместимости с разматывателями.",
-		questionCount: 6,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-21-1", question: "Какой внутренний диаметр втулки требуется?", answer: null },
-			{ id: "sq-21-2", question: "Допустим ли материал картон или нужен пластик?", answer: null },
-			{ id: "sq-21-3", question: "Максимальная масса рулона?", answer: null },
-			{ id: "sq-21-4", question: "Максимальный диаметр намотки?", answer: null },
-			{ id: "sq-21-5", question: "Какой допуск по ширине рулона?", answer: null },
-			{ id: "sq-21-6", question: "Возвратная ли втулка?", answer: null },
-		],
-		updatedAt: "2026-04-18T14:50:00.000Z",
-	},
-	{
-		id: "task-22",
-		name: "Реквизиты для выставления счёта",
-		status: "assigned",
-		item: ITEM_FABRIC,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-19T08:45:00.000Z",
-		deadlineAt: "2026-05-03T18:00:00.000Z",
-		description: "Поставщики просят актуальные реквизиты для предоставления счёта на предоплату.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-22-1", question: "Какое юрлицо выступает покупателем?", answer: null },
-			{ id: "sq-22-2", question: "Какой КПП указывать в счёте?", answer: null },
-			{ id: "sq-22-3", question: "Есть ли отдельная карточка предприятия?", answer: null },
-		],
-		updatedAt: "2026-04-19T08:45:00.000Z",
-	},
-	{
-		id: "task-23",
-		name: "Тендерная процедура и дорожная карта (ООО «Новые Материалы»)",
-		status: "assigned",
-		item: ITEM_LDSP,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-12T16:30:00.000Z",
-		deadlineAt: "2026-04-26T18:00:00.000Z",
-		description: "Поставщик просит детально описать процедуру выбора: этапы, критерии, сроки.",
-		questionCount: 8,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-23-1", question: "Сколько этапов отбора предусмотрено?", answer: null },
-			{ id: "sq-23-2", question: "Каков вес цены в итоговой оценке?", answer: null },
-			{ id: "sq-23-3", question: "Учитывается ли удалённость склада?", answer: null },
-			{ id: "sq-23-4", question: "Есть ли аудит производства?", answer: null },
-			{ id: "sq-23-5", question: "Проводится ли визит к поставщику?", answer: null },
-			{ id: "sq-23-6", question: "Когда объявляется победитель?", answer: null },
-			{ id: "sq-23-7", question: "Кто входит в тендерную комиссию?", answer: null },
-			{ id: "sq-23-8", question: "Публикуется ли протокол?", answer: null },
-		],
-		updatedAt: "2026-04-12T16:30:00.000Z",
-	},
-	{
-		id: "task-24",
-		name: "Возможность частичной отгрузки",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-20T10:15:00.000Z",
-		deadlineAt: "2026-05-04T18:00:00.000Z",
-		description: "Партнёр уточняет, готовы ли принимать плёнку частями, чтобы разгрузить производство.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-24-1", question: "Допустимо ли делить партию 15 000 м на 2 отгрузки?", answer: null },
-		],
-		updatedAt: "2026-04-20T10:15:00.000Z",
-	},
-	{
-		id: "task-25",
-		name: "Транспорт: фура 20 т или газель",
-		status: "assigned",
-		item: ITEM_PPU,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-11T09:00:00.000Z",
-		deadlineAt: "2026-04-25T18:00:00.000Z",
-		description: "Согласовать тип транспорта — нужны ли фуры или подойдёт малотоннажный транспорт.",
-		questionCount: 4,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-25-1", question: "Поместится ли партия в 3-тонную газель?", answer: null },
-			{ id: "sq-25-2", question: "Есть ли у поставщика своя логистика?", answer: null },
-			{ id: "sq-25-3", question: "Кто отвечает за выгрузку на нашем складе?", answer: null },
-			{ id: "sq-25-4", question: "Принимаете ли доставку сборным грузом?", answer: null },
-		],
-		updatedAt: "2026-04-11T09:00:00.000Z",
-	},
-	{
-		id: "task-26",
-		name: "Полная анкета для внесения в реестр поставщиков",
-		status: "assigned",
-		item: ITEM_LDSP,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-09T14:00:00.000Z",
-		deadlineAt: "2026-04-23T18:00:00.000Z",
-		description: "Для включения в реестр поставщиков нужно собрать расширенную анкету.",
-		questionCount: 9,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-26-1", question: "Год основания компании?", answer: null },
-			{ id: "sq-26-2", question: "Среднесписочная численность сотрудников?", answer: null },
-			{ id: "sq-26-3", question: "Годовой оборот за прошлый год?", answer: null },
-			{ id: "sq-26-4", question: "Основные клиенты в отрасли?", answer: null },
-			{ id: "sq-26-5", question: "Есть ли сертификация системы менеджмента?", answer: null },
-			{ id: "sq-26-6", question: "Судебные иски за последние 3 года?", answer: null },
-			{ id: "sq-26-7", question: "Бенефициары и доли участия?", answer: null },
-			{ id: "sq-26-8", question: "Готовы ли к аудиту финансовой отчётности?", answer: null },
-			{ id: "sq-26-9", question: "Производство собственное или подряд?", answer: null },
-		],
-		updatedAt: "2026-04-09T14:00:00.000Z",
-	},
-	{
-		id: "task-27",
-		name: "Подтверждение наличия плёнки на складе",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-21T12:40:00.000Z",
-		deadlineAt: "2026-05-05T18:00:00.000Z",
-		description: "Ключевой критерий — готовность отгрузить первую партию в течение недели.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-27-1", question: "Есть ли 15 000 м плёнки в наличии сейчас?", answer: null },
-			{ id: "sq-27-2", question: "Как быстро можно отгрузить после оплаты?", answer: null },
-		],
-		updatedAt: "2026-04-21T12:40:00.000Z",
-	},
-	{
-		id: "task-28",
-		name: "Возможность индивидуального фасона намотки",
-		status: "assigned",
-		item: ITEM_PVD,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-22T09:50:00.000Z",
-		deadlineAt: "2026-05-06T18:00:00.000Z",
-		description: "Нужна ли возможность заказа нестандартной длины рулона под конкретную линию упаковки.",
-		questionCount: 5,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-28-1", question: "Можно ли заказать рулон длиной 500 м вместо 1000 м?", answer: null },
-			{ id: "sq-28-2", question: "Влияет ли это на стоимость?", answer: null },
-			{ id: "sq-28-3", question: "Какая минимальная длина допустима?", answer: null },
-			{ id: "sq-28-4", question: "Сколько времени занимает перенастройка линии?", answer: null },
-			{ id: "sq-28-5", question: "Нужна ли предоплата за нестандарт?", answer: null },
-		],
-		updatedAt: "2026-04-22T09:50:00.000Z",
-	},
-	// --- In progress (13) ---
-	{
-		id: "task-29",
-		name: "Визит на производство (ООО «Пласт-Инвест»)",
-		status: "in_progress",
-		item: ITEM_PPU,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-05T11:30:00.000Z",
-		deadlineAt: "2026-04-26T18:00:00.000Z",
-		description: "Организуем визит на производство для финального отбора поставщика.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-29-1", question: "Удобная дата для визита?", answer: "15 мая 2026" },
-			{ id: "sq-29-2", question: "Сколько человек от нас принимаете?", answer: "До 3 человек" },
-			{ id: "sq-29-3", question: "Нужна ли предварительная заявка на пропуск?", answer: null },
-		],
-		updatedAt: "2026-04-20T14:00:00.000Z",
-	},
-	{
-		id: "task-30",
-		name: "Полный технический due-diligence поставщика (КРЕДО)",
-		status: "in_progress",
-		item: ITEM_PVD,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-02T10:20:00.000Z",
-		deadlineAt: "2026-04-30T18:00:00.000Z",
-		description: "Глубокий технический анализ: оборудование, контроль качества, резервы.",
-		questionCount: 10,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-30-1", question: "Какое оборудование используется для экструзии?", answer: "Reifenhäuser 2019" },
-			{ id: "sq-30-2", question: "Сколько линий в работе?", answer: "3 линии" },
-			{ id: "sq-30-3", question: "Какая общая мощность в месяц?", answer: "250 тонн" },
-			{ id: "sq-30-4", question: "Как проводится входной контроль сырья?", answer: "Лабораторный анализ партии" },
-			{ id: "sq-30-5", question: "Есть ли резервная линия?", answer: null },
-			{ id: "sq-30-6", question: "Частота калибровки оборудования?", answer: null },
-			{ id: "sq-30-7", question: "Контроль толщины онлайн или выборочно?", answer: null },
-			{ id: "sq-30-8", question: "Какой процент брака считаете нормой?", answer: null },
-			{ id: "sq-30-9", question: "Есть ли архив параметров каждой партии?", answer: null },
-			{ id: "sq-30-10", question: "Каким образом утилизируется брак?", answer: null },
-		],
-		updatedAt: "2026-04-19T16:45:00.000Z",
-	},
-	{
-		id: "task-31",
-		name: "Согласование графика отгрузок на Q2",
-		status: "in_progress",
-		item: ITEM_PPU,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-07T09:00:00.000Z",
-		deadlineAt: "2026-04-28T18:00:00.000Z",
-		description: "Составляем график отгрузок с поставщиком на второй квартал.",
-		questionCount: 6,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-31-1", question: "Отгрузка в апреле — какая неделя?", answer: "17 неделя" },
-			{ id: "sq-31-2", question: "Май — возможно ли две отгрузки?", answer: "Да" },
-			{ id: "sq-31-3", question: "Июнь — учитываем праздники?", answer: "Да, перенос на 15 июня" },
-			{ id: "sq-31-4", question: "Как уведомляется о готовности?", answer: null },
-			{ id: "sq-31-5", question: "Можно ли сместить отгрузку за 3 дня?", answer: null },
-			{ id: "sq-31-6", question: "Штрафы за срыв сроков?", answer: null },
-		],
-		updatedAt: "2026-04-21T10:20:00.000Z",
-	},
-	{
-		id: "task-32",
-		name: "Подписание NDA перед обменом спецификациями",
-		status: "in_progress",
-		item: ITEM_FABRIC,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-10T13:45:00.000Z",
-		deadlineAt: "2026-04-24T18:00:00.000Z",
-		description: "Перед передачей детальной спецификации подписываем соглашение о неразглашении.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-32-1", question: "Готовы ли подписать типовое NDA ОРМАТЕК?", answer: "Да" },
-			{ id: "sq-32-2", question: "Нужны ли правки в договоре?", answer: null },
-		],
-		updatedAt: "2026-04-18T11:30:00.000Z",
-	},
-	{
-		id: "task-33",
-		name: "Детальная калькуляция TCO (7 поставщиков)",
-		status: "in_progress",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-01T15:00:00.000Z",
-		deadlineAt: "2026-04-27T18:00:00.000Z",
-		description: "Считаем полную стоимость владения: цена, доставка, хранение, отсрочка, потери.",
-		questionCount: 7,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-33-1", question: "Стоимость доставки отдельно или в цене?", answer: "Отдельно" },
-			{ id: "sq-33-2", question: "Дни отсрочки фактические?", answer: "7 дней" },
-			{ id: "sq-33-3", question: "Процент возвратов по прошлым сделкам?", answer: "0,3%" },
-			{ id: "sq-33-4", question: "Средняя потеря при раскатке?", answer: null },
-			{ id: "sq-33-5", question: "Страховка груза включена?", answer: null },
-			{ id: "sq-33-6", question: "Выгрузка силами поставщика?", answer: null },
-			{ id: "sq-33-7", question: "Стоимость хранения у поставщика до 14 дней?", answer: null },
-		],
-		updatedAt: "2026-04-22T13:50:00.000Z",
-	},
-	{
-		id: "task-34",
-		name: "Пересогласование КП после изменения курса валют",
-		status: "in_progress",
-		item: ITEM_PPU,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-14T10:10:00.000Z",
-		deadlineAt: "2026-04-28T18:00:00.000Z",
-		description: "После роста курса поставщик запросил пересмотр цены на импортное сырьё.",
-		questionCount: 4,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-34-1", question: "На сколько процентов предлагаете повысить цену?", answer: "7%" },
-			{ id: "sq-34-2", question: "Привязка к какому курсу?", answer: "ЦБ РФ" },
-			{ id: "sq-34-3", question: "Возможна ли фиксация коридора ±3%?", answer: null },
-			{ id: "sq-34-4", question: "Как часто пересматриваем цену?", answer: null },
-		],
-		updatedAt: "2026-04-21T09:00:00.000Z",
-	},
-	{
-		id: "task-35",
-		name: "Подписание договора (финальная правка)",
-		status: "in_progress",
-		item: ITEM_FABRIC,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-08T16:15:00.000Z",
-		deadlineAt: "2026-04-25T18:00:00.000Z",
-		description: "Юристы обменялись правками — согласовываем финальную редакцию.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [{ id: "sq-35-1", question: "Принимаете ли последние правки пункта 7.2?", answer: null }],
-		updatedAt: "2026-04-20T17:00:00.000Z",
-	},
-	{
-		id: "task-36",
-		name: "Тестовая партия 500 м (Polystan)",
-		status: "in_progress",
-		item: ITEM_PVD,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-06T11:00:00.000Z",
-		deadlineAt: "2026-04-26T18:00:00.000Z",
-		description: "Запускаем тестовую партию для проверки в производстве перед основным контрактом.",
-		questionCount: 5,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-36-1", question: "Сроки производства тестовой партии?", answer: "7 рабочих дней" },
-			{ id: "sq-36-2", question: "Стоимость тестовой партии?", answer: "По прайсу" },
-			{ id: "sq-36-3", question: "Возможен ли возврат при несоответствии?", answer: null },
-			{ id: "sq-36-4", question: "Как оформлять приёмку?", answer: null },
-			{ id: "sq-36-5", question: "Сколько времени на тестирование у нас?", answer: null },
-		],
-		updatedAt: "2026-04-19T12:30:00.000Z",
-	},
-	{
-		id: "task-37",
-		name: "Юридический аудит поставщика (Сталерпласт)",
-		status: "in_progress",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-03T14:25:00.000Z",
-		deadlineAt: "2026-04-29T18:00:00.000Z",
-		description: "Проверка юрлица: учредительные документы, лицензии, отсутствие обременений.",
-		questionCount: 8,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-37-1", question: "Предоставите ли ЕГРЮЛ свежей датой?", answer: "Да" },
-			{ id: "sq-37-2", question: "Выписку по банковскому счёту?", answer: "Да" },
-			{ id: "sq-37-3", question: "Копию устава?", answer: "Да" },
-			{ id: "sq-37-4", question: "Решение о назначении директора?", answer: null },
-			{ id: "sq-37-5", question: "Справку об отсутствии задолженности?", answer: null },
-			{ id: "sq-37-6", question: "Лицензии на производственную деятельность?", answer: null },
-			{ id: "sq-37-7", question: "Страховка ответственности?", answer: null },
-			{ id: "sq-37-8", question: "Бухгалтерская отчётность за 2 года?", answer: null },
-		],
-		updatedAt: "2026-04-22T10:15:00.000Z",
-	},
-	{
-		id: "task-38",
-		name: "Подготовка договора поставки (шаблон ОРМАТЕК)",
-		status: "in_progress",
-		item: ITEM_FABRIC,
-		assignee: assigneeOlga,
-		createdAt: "2026-04-11T08:30:00.000Z",
-		deadlineAt: "2026-04-25T18:00:00.000Z",
-		description: "Направили шаблон договора поставщику — ждём акцепта или правок.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-38-1", question: "Есть ли комментарии к пункту о штрафах?", answer: "Да, предложим свою редакцию" },
-			{ id: "sq-38-2", question: "Согласны ли с подсудностью в Ростове?", answer: null },
-			{ id: "sq-38-3", question: "Нужны ли дополнительные приложения?", answer: null },
-		],
-		updatedAt: "2026-04-21T15:00:00.000Z",
-	},
-	{
-		id: "task-39",
-		name: "Уточнение параметров сырья (Polimer-Group)",
-		status: "in_progress",
-		item: ITEM_PVD,
-		assignee: assigneeDmitry,
-		createdAt: "2026-04-04T12:00:00.000Z",
-		deadlineAt: "2026-04-27T18:00:00.000Z",
-		description: "Поставщик предоставил паспорт сырья — уточняем ключевые параметры.",
-		questionCount: 6,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-39-1", question: "Индекс текучести расплава?", answer: "2,0 г/10 мин" },
-			{ id: "sq-39-2", question: "Плотность сырья?", answer: "0,92 г/см³" },
-			{ id: "sq-39-3", question: "Добавки и их процент?", answer: "0,5% стабилизатор" },
-			{ id: "sq-39-4", question: "Производитель сырья?", answer: null },
-			{ id: "sq-39-5", question: "Регион производства сырья?", answer: null },
-			{ id: "sq-39-6", question: "Альтернативный поставщик сырья?", answer: null },
-		],
-		updatedAt: "2026-04-20T11:40:00.000Z",
-	},
-	{
-		id: "task-40",
-		name: "Комплексная проверка перед включением в шорт-лист",
-		status: "in_progress",
-		item: ITEM_LDSP,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-04-02T09:40:00.000Z",
-		deadlineAt: "2026-04-30T18:00:00.000Z",
-		description: "Финальный чек-лист перед переходом к переговорам по цене.",
-		questionCount: 9,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-40-1", question: "Производство собственное?", answer: "Да" },
-			{ id: "sq-40-2", question: "Опыт работы с федеральными клиентами?", answer: "Да, 3 года" },
-			{ id: "sq-40-3", question: "Готовность к аудиту качества?", answer: "Да" },
-			{ id: "sq-40-4", question: "Стабильность цен последний год?", answer: null },
-			{ id: "sq-40-5", question: "Наличие резервов мощностей?", answer: null },
-			{ id: "sq-40-6", question: "Готовность к ЭДО?", answer: null },
-			{ id: "sq-40-7", question: "Страховка ответственности?", answer: null },
-			{ id: "sq-40-8", question: "История задержек отгрузок?", answer: null },
-			{ id: "sq-40-9", question: "Работа с нашими основными конкурентами?", answer: null },
-		],
-		updatedAt: "2026-04-22T08:50:00.000Z",
-	},
-	{
-		id: "task-41",
-		name: "Формат счёта-фактуры в Диадоке",
-		status: "in_progress",
-		item: ITEM_FABRIC,
-		assignee: assigneeIvan,
-		createdAt: "2026-04-13T10:00:00.000Z",
-		deadlineAt: "2026-04-25T18:00:00.000Z",
-		description: "Уточняем формат обмена УПД в электронном виде.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-41-1", question: "Используете ли УПД-1 или УПД-2?", answer: "УПД-2" },
-			{ id: "sq-41-2", question: "Готовы ли к роумингу между операторами?", answer: null },
-		],
-		updatedAt: "2026-04-21T16:20:00.000Z",
-	},
-	// --- Completed (15 more) ---
-	{
-		id: "task-42",
-		name: "Проверка упаковки рулонов в термоусадку",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeOlga,
-		createdAt: "2026-03-01T10:00:00.000Z",
-		deadlineAt: "2026-03-08T18:00:00.000Z",
-		description: "Нужна упаковка рулонов в термоусадку для защиты при транспортировке.",
-		questionCount: 4,
-		completedResponse: "Все рулоны упаковываются в термоусадку + палетирование. Включено в цену.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-42-1", question: "Каждый ли рулон в термоусадке?", answer: "Да" },
-			{ id: "sq-42-2", question: "Палетируется ли груз?", answer: "Да, по 4 рулона на палете" },
-			{ id: "sq-42-3", question: "Дополнительная стоимость упаковки?", answer: "Включено в цену" },
-			{ id: "sq-42-4", question: "Маркировка палеты?", answer: "Да, артикул + партия" },
-		],
-		updatedAt: "2026-03-03T11:00:00.000Z",
-	},
-	{
-		id: "task-43",
-		name: "Срок годности плёнки",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeDmitry,
-		createdAt: "2026-03-02T14:00:00.000Z",
-		deadlineAt: "2026-03-09T18:00:00.000Z",
-		description: "Уточнить гарантийный срок хранения плёнки.",
-		questionCount: 1,
-		completedResponse: "Срок годности плёнки ПВД — 24 месяца при соблюдении условий хранения.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [{ id: "sq-43-1", question: "Какой гарантийный срок хранения?", answer: "24 месяца" }],
-		updatedAt: "2026-03-04T09:30:00.000Z",
-	},
-	{
-		id: "task-44",
-		name: "Условия рекламации и возврата",
-		status: "completed",
-		item: ITEM_PPU,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-03-05T11:20:00.000Z",
-		deadlineAt: "2026-03-12T18:00:00.000Z",
-		description: "Согласовать процедуру рекламации при несоответствии качества.",
-		questionCount: 6,
-		completedResponse:
-			"Рекламация принимается в течение 10 дней с момента приёмки. Замена или возврат денег — по выбору ОРМАТЕК.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-44-1", question: "Срок подачи рекламации?", answer: "10 календарных дней" },
-			{ id: "sq-44-2", question: "Формат фиксации брака?", answer: "Акт + фотоотчёт" },
-			{ id: "sq-44-3", question: "Компенсация — замена или деньги?", answer: "По выбору покупателя" },
-			{ id: "sq-44-4", question: "Кто оплачивает возврат?", answer: "Поставщик, если брак подтверждён" },
-			{ id: "sq-44-5", question: "Срок рассмотрения?", answer: "5 рабочих дней" },
-			{ id: "sq-44-6", question: "Участие независимой экспертизы?", answer: "По согласованию" },
-		],
-		updatedAt: "2026-03-08T15:00:00.000Z",
-	},
-	{
-		id: "task-45",
-		name: "Выбор инкотермса для поставки",
-		status: "completed",
-		item: ITEM_FABRIC,
-		assignee: assigneeIvan,
-		createdAt: "2026-03-06T13:00:00.000Z",
-		deadlineAt: "2026-03-13T18:00:00.000Z",
-		description: "Выбираем базис поставки по Incoterms 2020.",
-		questionCount: 3,
-		completedResponse: "Договорились на DAP до склада в Аксайском районе. Переход риска — при выгрузке.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-45-1", question: "Какой базис удобен?", answer: "DAP" },
-			{ id: "sq-45-2", question: "Точка перехода рисков?", answer: "При выгрузке на нашем складе" },
-			{ id: "sq-45-3", question: "Страхование груза?", answer: "За счёт поставщика" },
-		],
-		updatedAt: "2026-03-09T10:00:00.000Z",
-	},
-	{
-		id: "task-46",
-		name: "Проверка качества первичного сырья (лабораторные параметры)",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeOlga,
-		createdAt: "2026-03-07T09:15:00.000Z",
-		deadlineAt: "2026-03-17T18:00:00.000Z",
-		description: "Согласовали расширенный лабораторный контроль.",
-		questionCount: 8,
-		completedResponse:
-			"Поставщик предоставляет полный протокол ИК-спектроскопии + механические испытания по каждой партии.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-46-1", question: "ИК-спектроскопия проводится?", answer: "Да, по каждой партии" },
-			{ id: "sq-46-2", question: "Прочность на разрыв?", answer: "≥18 МПа" },
-			{ id: "sq-46-3", question: "Удлинение при разрыве?", answer: "≥600%" },
-			{ id: "sq-46-4", question: "Тест на разнотолщинность?", answer: "Да, ±5%" },
-			{ id: "sq-46-5", question: "Определение вторсырья?", answer: "Да, по ИК" },
-			{ id: "sq-46-6", question: "Частота контроля?", answer: "Каждая партия" },
-			{ id: "sq-46-7", question: "Срок получения протокола?", answer: "3 рабочих дня" },
-			{ id: "sq-46-8", question: "Протокол на русском?", answer: "Да" },
-		],
-		updatedAt: "2026-03-12T14:30:00.000Z",
-	},
-	{
-		id: "task-47",
-		name: "Сертификат СЭЗ на плёнку для пищевого назначения",
-		status: "completed",
-		item: ITEM_PVD,
-		assignee: assigneeDmitry,
-		createdAt: "2026-03-09T14:45:00.000Z",
-		deadlineAt: "2026-03-16T18:00:00.000Z",
-		description: "Проверяем допуск плёнки к контакту с пищевыми продуктами.",
-		questionCount: 2,
-		completedResponse: "Плёнка имеет СЭЗ Роспотребнадзора, допущена к контакту с пищевыми продуктами.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-47-1", question: "Есть ли СЭЗ?", answer: "Да, от 15.01.2025" },
-			{ id: "sq-47-2", question: "Область применения?", answer: "Контакт с сухими пищевыми продуктами" },
-		],
-		updatedAt: "2026-03-11T09:00:00.000Z",
-	},
-	{
-		id: "task-48",
-		name: "Система скидок при росте объёмов",
-		status: "completed",
-		item: ITEM_PPU,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-03-10T11:00:00.000Z",
-		deadlineAt: "2026-03-17T18:00:00.000Z",
-		description: "Согласуем прогрессивную шкалу скидок.",
-		questionCount: 5,
-		completedResponse: "Скидки: от 20 т/мес — 2%, от 30 т — 4%, от 50 т — 6%. Пересмотр каждые 3 месяца.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-48-1", question: "Скидка от 20 т в месяц?", answer: "2%" },
-			{ id: "sq-48-2", question: "От 30 т?", answer: "4%" },
-			{ id: "sq-48-3", question: "От 50 т?", answer: "6%" },
-			{ id: "sq-48-4", question: "Период пересмотра?", answer: "Каждые 3 месяца" },
-			{ id: "sq-48-5", question: "Фиксация в договоре?", answer: "Приложение №2" },
-		],
-		updatedAt: "2026-03-13T16:20:00.000Z",
-	},
-	{
-		id: "task-49",
-		name: "Комплексные условия стратегического партнёрства",
-		status: "completed",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-03-11T08:30:00.000Z",
-		deadlineAt: "2026-03-25T18:00:00.000Z",
-		description: "Обсуждаем долгосрочный контракт на 2 года с резервированием мощностей.",
-		questionCount: 10,
-		completedResponse:
-			"Подписан рамочный договор на 24 месяца: резерв мощности 30 т/мес, скидка 5%, приоритет отгрузки, ежеквартальные встречи.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-49-1", question: "Срок контракта?", answer: "24 месяца" },
-			{ id: "sq-49-2", question: "Резерв мощности?", answer: "30 т/мес" },
-			{ id: "sq-49-3", question: "Стратегическая скидка?", answer: "5% к прайсу" },
-			{ id: "sq-49-4", question: "Приоритет в очереди отгрузки?", answer: "Да" },
-			{ id: "sq-49-5", question: "Эксклюзивность по региону?", answer: "Нет" },
-			{ id: "sq-49-6", question: "Квартальные встречи?", answer: "Да, онлайн" },
-			{ id: "sq-49-7", question: "Разработка спецвидов плёнки?", answer: "По запросу" },
-			{ id: "sq-49-8", question: "Совместный R&D бюджет?", answer: "Нет" },
-			{ id: "sq-49-9", question: "Условия досрочного расторжения?", answer: "За 90 дней уведомления" },
-			{ id: "sq-49-10", question: "Индексация цены?", answer: "Раз в 6 мес по ИПЦ" },
-		],
-		updatedAt: "2026-03-18T13:00:00.000Z",
-	},
-	{
-		id: "task-50",
-		name: "Запрос истории отгрузок за последний год",
-		status: "completed",
-		item: ITEM_PPU,
-		assignee: assigneeOlga,
-		createdAt: "2026-03-12T10:30:00.000Z",
-		deadlineAt: "2026-03-19T18:00:00.000Z",
-		description: "Нужна статистика: сколько тонн поставлено, процент брака, задержки.",
-		questionCount: 7,
-		completedResponse: "Получен отчёт: 180 т, брак 0,4%, средняя задержка 1 день. Результат приемлемый.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-50-1", question: "Общий объём поставок за год?", answer: "180 тонн" },
-			{ id: "sq-50-2", question: "Процент брака?", answer: "0,4%" },
-			{ id: "sq-50-3", question: "Среднее отклонение от срока?", answer: "1 день" },
-			{ id: "sq-50-4", question: "Количество уникальных клиентов?", answer: "12" },
-			{ id: "sq-50-5", question: "Крупнейший клиент?", answer: "ООО «Упак-М»" },
-			{ id: "sq-50-6", question: "Отказов от поставок?", answer: "1 случай" },
-			{ id: "sq-50-7", question: "Сезонные колебания?", answer: "Лето: −20%" },
-		],
-		updatedAt: "2026-03-15T14:00:00.000Z",
-	},
-	{
-		id: "task-51",
-		name: "Пеня за просрочку платежа",
-		status: "completed",
-		item: ITEM_LDSP,
-		assignee: assigneeDmitry,
-		createdAt: "2026-03-14T15:00:00.000Z",
-		deadlineAt: "2026-03-21T18:00:00.000Z",
-		description: "Согласовать размер пени в договоре.",
-		questionCount: 3,
-		completedResponse: "Пеня — 0,1% в день, но не более 10% от суммы просроченного платежа.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-51-1", question: "Процент пени?", answer: "0,1% в день" },
-			{ id: "sq-51-2", question: "Максимум пени?", answer: "10% от суммы" },
-			{ id: "sq-51-3", question: "Грейс-период?", answer: "3 рабочих дня" },
-		],
-		updatedAt: "2026-03-17T10:00:00.000Z",
-	},
-	{
-		id: "task-52",
-		name: "Включение ОРМАТЕК в шорт-лист для тестирования новых марок",
-		status: "completed",
-		item: ITEM_PPU,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-03-15T11:45:00.000Z",
-		deadlineAt: "2026-03-22T18:00:00.000Z",
-		description: "Поставщик предлагает участвовать в пилотах новых марок плёнки.",
-		questionCount: 4,
-		completedResponse: "Включены в шорт-лист. Первый пилот — плёнка с улучшенной прочностью в мае.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-52-1", question: "Готовы ли участвовать в пилоте?", answer: "Да" },
-			{ id: "sq-52-2", question: "Бесплатная ли тестовая партия?", answer: "Да, до 100 м" },
-			{ id: "sq-52-3", question: "Обратная связь — формат?", answer: "Отчёт + встреча" },
-			{ id: "sq-52-4", question: "Эксклюзивность теста?", answer: "Нет" },
-		],
-		updatedAt: "2026-03-19T09:30:00.000Z",
-	},
-	{
-		id: "task-53",
-		name: "Контактное лицо по оперативным вопросам",
-		status: "completed",
-		item: ITEM_FABRIC,
-		assignee: assigneeIvan,
-		createdAt: "2026-03-16T09:20:00.000Z",
-		deadlineAt: "2026-03-20T18:00:00.000Z",
-		description: "Уточнить, кто отвечает за срочные запросы со стороны поставщика.",
-		questionCount: 1,
-		completedResponse: "Контакт: Елена Жукова, +7 (812) 555-00-12, e.zhukova@polystar.ru. Чат в WhatsApp.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [{ id: "sq-53-1", question: "Кто ответственный менеджер?", answer: "Елена Жукова" }],
-		updatedAt: "2026-03-17T11:00:00.000Z",
-	},
-	{
-		id: "task-54",
-		name: "Квартальная отчётность по контракту",
-		status: "completed",
-		item: ITEM_LDSP,
-		assignee: assigneeOlga,
-		createdAt: "2026-03-17T13:00:00.000Z",
-		deadlineAt: "2026-03-24T18:00:00.000Z",
-		description: "Согласовать формат и содержание квартального отчёта от поставщика.",
-		questionCount: 5,
-		completedResponse: "Отчёт — Excel + PDF summary, предоставляется в течение 5 рабочих дней после закрытия квартала.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-54-1", question: "Формат отчёта?", answer: "Excel + PDF" },
-			{ id: "sq-54-2", question: "Срок предоставления?", answer: "5 рабочих дней" },
-			{ id: "sq-54-3", question: "Разделы отчёта?", answer: "Объёмы, задержки, брак, NPS" },
-			{ id: "sq-54-4", question: "Презентация отчёта?", answer: "Онлайн-встреча" },
-			{ id: "sq-54-5", question: "Участники встречи?", answer: "Ком. директор + снабжение" },
-		],
-		updatedAt: "2026-03-21T16:00:00.000Z",
-	},
-	{
-		id: "task-55",
-		name: "Резервный поставщик на случай форс-мажора",
-		status: "completed",
-		item: ITEM_PPU,
-		assignee: assigneeDmitry,
-		createdAt: "2026-03-18T10:00:00.000Z",
-		deadlineAt: "2026-03-28T18:00:00.000Z",
-		description: "Определяем стратегию работы при невозможности отгрузки основным поставщиком.",
-		questionCount: 6,
-		completedResponse:
-			"В качестве резерва утверждён ООО «Пластик-Сервис» с обязательством отгрузить 5 т в течение 72 часов.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-55-1", question: "Кто резервный поставщик?", answer: "ООО «Пластик-Сервис»" },
-			{ id: "sq-55-2", question: "Срок активации?", answer: "72 часа" },
-			{ id: "sq-55-3", question: "Объём резерва?", answer: "5 тонн" },
-			{ id: "sq-55-4", question: "Цена относительно основной?", answer: "+12%" },
-			{ id: "sq-55-5", question: "Платёж за резервирование?", answer: "Нет" },
-			{ id: "sq-55-6", question: "Пересмотр списка?", answer: "Раз в 6 мес" },
-		],
-		updatedAt: "2026-03-24T11:15:00.000Z",
-	},
-	{
-		id: "task-56",
-		name: "Переход на двойной контроль отгрузочных документов",
-		status: "completed",
-		item: ITEM_FABRIC,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-03-20T14:30:00.000Z",
-		deadlineAt: "2026-03-27T18:00:00.000Z",
-		description: "Чтобы избежать ошибок в ТТН, внедряем двойной контроль у поставщика.",
-		questionCount: 2,
-		completedResponse: "Поставщик внедрил двойной контроль. Ошибки в ТТН снизились до 0.",
-		attachments: [],
-		statusBeforeArchive: null,
-		supplierQuestions: [
-			{ id: "sq-56-1", question: "Готовы ли внедрить двойной контроль?", answer: "Да" },
-			{ id: "sq-56-2", question: "Срок внедрения?", answer: "2 недели" },
-		],
-		updatedAt: "2026-03-23T10:45:00.000Z",
-	},
-	// --- Archived (10) ---
-	{
-		id: "task-57",
-		name: "Поставка плёнки через посредника (отклонено)",
-		status: "archived",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-05T10:00:00.000Z",
-		deadlineAt: "2026-02-12T18:00:00.000Z",
-		description: "Посредник предлагает схему со своей маржой — нерентабельно.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "assigned",
-		supplierQuestions: [
-			{ id: "sq-57-1", question: "Какая маржа посредника?", answer: "12%" },
-			{ id: "sq-57-2", question: "Можно ли напрямую с заводом?", answer: "Нет, работаем через нас" },
-			{ id: "sq-57-3", question: "Готовы ли на 5%?", answer: "Нет" },
-		],
-		updatedAt: "2026-02-10T16:00:00.000Z",
-	},
-	{
-		id: "task-58",
-		name: "Запрос на вторсырьё (отклонено по политике)",
-		status: "archived",
-		item: ITEM_PVD,
-		assignee: assigneeOlga,
-		createdAt: "2026-02-08T11:30:00.000Z",
-		deadlineAt: "2026-02-15T18:00:00.000Z",
-		description: "Поставщик предлагает микс с вторсырьём. Политика — только первичное.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "assigned",
-		supplierQuestions: [{ id: "sq-58-1", question: "Рассматриваем ли 70/30 первичка/вторсырьё?", answer: "Нет" }],
-		updatedAt: "2026-02-11T13:20:00.000Z",
-	},
-	{
-		id: "task-59",
-		name: "Неподтверждённые мощности (Polyprom)",
-		status: "archived",
-		item: ITEM_PPU,
-		assignee: assigneeDmitry,
-		createdAt: "2026-02-10T09:00:00.000Z",
-		deadlineAt: "2026-02-17T18:00:00.000Z",
-		description: "Поставщик не смог подтвердить заявленные мощности производства.",
-		questionCount: 4,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "in_progress",
-		supplierQuestions: [
-			{ id: "sq-59-1", question: "Реальная мощность в месяц?", answer: "Не подтверждена" },
-			{ id: "sq-59-2", question: "Количество действующих линий?", answer: "1 вместо 3" },
-			{ id: "sq-59-3", question: "Возможен ли аудит?", answer: "Отказались" },
-			{ id: "sq-59-4", question: "Есть ли резервные мощности?", answer: "Нет" },
-		],
-		updatedAt: "2026-02-16T11:00:00.000Z",
-	},
-	{
-		id: "task-60",
-		name: "Несоответствие сертификации (ООО «БыстроПласт»)",
-		status: "archived",
-		item: ITEM_LDSP,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-02-12T14:00:00.000Z",
-		deadlineAt: "2026-02-19T18:00:00.000Z",
-		description: "У поставщика отсутствует актуальный СЭЗ, сертификат просрочен.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "assigned",
-		supplierQuestions: [
-			{ id: "sq-60-1", question: "Дата окончания СЭЗ?", answer: "Истёк 11.2025" },
-			{ id: "sq-60-2", question: "Когда получат новый?", answer: "Не ранее Q3 2026" },
-		],
-		updatedAt: "2026-02-14T10:30:00.000Z",
-	},
-	{
-		id: "task-61",
-		name: "Слишком высокая цена (СтройПак)",
-		status: "archived",
-		item: ITEM_LDSP,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-14T13:00:00.000Z",
-		deadlineAt: "2026-02-21T18:00:00.000Z",
-		description: "Цена выше рынка на 25%, переговоры нерентабельны.",
-		questionCount: 5,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "in_progress",
-		supplierQuestions: [
-			{ id: "sq-61-1", question: "Почему цена выше рынка?", answer: "Премиум-сегмент" },
-			{ id: "sq-61-2", question: "Готовы ли снизить цену?", answer: "Нет" },
-			{ id: "sq-61-3", question: "Включены ли доп. услуги?", answer: "Нет" },
-			{ id: "sq-61-4", question: "Скидка при объёме?", answer: "Максимум 3%" },
-			{ id: "sq-61-5", question: "Готовность к переговорам?", answer: "Не готовы" },
-		],
-		updatedAt: "2026-02-18T15:00:00.000Z",
-	},
-	{
-		id: "task-62",
-		name: "Неподходящие условия оплаты (ПакМаркет)",
-		status: "archived",
-		item: ITEM_FABRIC,
-		assignee: assigneeOlga,
-		createdAt: "2026-02-16T10:20:00.000Z",
-		deadlineAt: "2026-02-23T18:00:00.000Z",
-		description: "Поставщик требует 100% предоплату на полгода вперёд.",
-		questionCount: 6,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "in_progress",
-		supplierQuestions: [
-			{ id: "sq-62-1", question: "Возможен ли аванс 50%?", answer: "Нет" },
-			{ id: "sq-62-2", question: "Помесячная оплата?", answer: "Нет" },
-			{ id: "sq-62-3", question: "Банковская гарантия?", answer: "Нет" },
-			{ id: "sq-62-4", question: "Предоплата на партию?", answer: "Нет" },
-			{ id: "sq-62-5", question: "Скидка за 100% предоплату?", answer: "2%" },
-			{ id: "sq-62-6", question: "Срок действия цены?", answer: "1 мес" },
-		],
-		updatedAt: "2026-02-20T09:00:00.000Z",
-	},
-	{
-		id: "task-63",
-		name: "Отказ от тендера (ПолиМир)",
-		status: "archived",
-		item: ITEM_LDSP,
-		assignee: assigneeDmitry,
-		createdAt: "2026-02-18T15:00:00.000Z",
-		deadlineAt: "2026-02-25T18:00:00.000Z",
-		description: "Поставщик отказался от участия в тендере по внутренним причинам.",
-		questionCount: 1,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "assigned",
-		supplierQuestions: [{ id: "sq-63-1", question: "Причина отказа?", answer: "Загрузка производства до Q3" }],
-		updatedAt: "2026-02-19T11:20:00.000Z",
-	},
-	{
-		id: "task-64",
-		name: "Провальное качество на тестовой партии (Росплёнка)",
-		status: "archived",
-		item: ITEM_PVD,
-		assignee: assigneeEkaterina,
-		createdAt: "2026-02-20T09:00:00.000Z",
-		deadlineAt: "2026-03-05T18:00:00.000Z",
-		description: "Тестовая партия показала разнотолщинность и вкрапления — снимаем с рассмотрения.",
-		questionCount: 8,
-		completedResponse:
-			"Тест не пройден. Разнотолщинность до 18% при норме 5%, вкрапления в 3 из 5 рулонов. Выдан отрицательный протокол.",
-		attachments: [],
-		statusBeforeArchive: "completed",
-		supplierQuestions: [
-			{ id: "sq-64-1", question: "Разнотолщинность тестовой партии?", answer: "18% (норма 5%)" },
-			{ id: "sq-64-2", question: "Количество рулонов с дефектами?", answer: "3 из 5" },
-			{ id: "sq-64-3", question: "Готовы ли на бесплатную замену?", answer: "Да, но не сейчас" },
-			{ id: "sq-64-4", question: "Причина брака?", answer: "Неисправность калибра" },
-			{ id: "sq-64-5", question: "Срок устранения?", answer: "2 месяца" },
-			{ id: "sq-64-6", question: "Готовы ли повторить тест?", answer: "Не раньше Q3" },
-			{ id: "sq-64-7", question: "Компенсация наших расходов на ОТК?", answer: "Нет" },
-			{ id: "sq-64-8", question: "Итог?", answer: "Архивируем" },
-		],
-		updatedAt: "2026-03-03T14:00:00.000Z",
-	},
-	{
-		id: "task-65",
-		name: "Конфликт интересов (связь с конкурентом)",
-		status: "archived",
-		item: ITEM_FABRIC,
-		assignee: assigneeIvan,
-		createdAt: "2026-02-22T12:00:00.000Z",
-		deadlineAt: "2026-02-27T18:00:00.000Z",
-		description: "Выяснилось, что поставщик — эксклюзивный партнёр нашего прямого конкурента.",
-		questionCount: 2,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "assigned",
-		supplierQuestions: [
-			{ id: "sq-65-1", question: "Подтверждаете эксклюзивность с конкурентом?", answer: "Да" },
-			{ id: "sq-65-2", question: "Возможна ли работа с нами?", answer: "Нет" },
-		],
-		updatedAt: "2026-02-23T13:15:00.000Z",
-	},
-	{
-		id: "task-66",
-		name: "Запрос отклонён по геополитическим рискам",
-		status: "archived",
-		item: ITEM_PPU,
-		assignee: assigneeOlga,
-		createdAt: "2026-02-25T10:30:00.000Z",
-		deadlineAt: "2026-03-04T18:00:00.000Z",
-		description: "Сырьё зависит от импорта из недружественной юрисдикции — риски слишком высокие.",
-		questionCount: 3,
-		completedResponse: null,
-		attachments: [],
-		statusBeforeArchive: "in_progress",
-		supplierQuestions: [
-			{ id: "sq-66-1", question: "Доля импортного сырья?", answer: "70%" },
-			{ id: "sq-66-2", question: "Планы по локализации?", answer: "Нет" },
-			{ id: "sq-66-3", question: "Альтернативные поставщики сырья?", answer: "В поиске" },
-		],
-		updatedAt: "2026-03-01T14:00:00.000Z",
-	},
-];
-
-// --- Mutable store ---
-
-let tasksStore: Task[] = [];
-
-function cloneTask(t: Task): Task {
-	return {
-		...t,
-		item: { ...t.item },
-		assignee: t.assignee ? { ...t.assignee } : null,
-		attachments: t.attachments.map((a) => ({ ...a })),
-		supplierQuestions: t.supplierQuestions.map((q) => ({ ...q })),
-	};
-}
-
-function hydrateSeedTasks(seed: SeedTask[]): Task[] {
-	const pool = ORMATEK_SUPPLIERS;
-	let cursor = 0;
-	return seed.map((task) => {
-		const taskCreatedMs = new Date(task.createdAt).getTime();
-		return {
-			...task,
-			supplierQuestions: task.supplierQuestions.map((q, qIdx) => {
-				const supplier = pool[cursor % pool.length];
-				cursor += 1;
-				const askedMs = taskCreatedMs - (qIdx + 1) * 24 * 3600 * 1000;
-				return {
-					...q,
-					supplierId: supplier.id,
-					supplierName: supplier.companyName,
-					askedAt: new Date(askedMs).toISOString(),
-				};
-			}),
-		};
-	});
-}
-
-function seedStore() {
-	tasksStore = hydrateSeedTasks(SEED_TASKS).map(cloneTask);
-}
-
-seedStore();
-
-export function _resetTasksStore(): void {
-	seedStore();
-}
-
-export function _setTasks(tasks: Task[]): void {
-	tasksStore = tasks.map(cloneTask);
-}
-
-export function _getAllTasks(): Task[] {
-	return tasksStore.map(cloneTask);
-}
-
-// --- Internal helpers ---
-
-function findTaskIndex(id: string): number {
-	return tasksStore.findIndex((t) => t.id === id);
-}
-
-function requireTaskIdx(id: string): number {
-	const idx = findTaskIndex(id);
-	if (idx === -1) throw new Error(`Task ${id} not found`);
-	return idx;
-}
-
-type TaskSortField = "created_at" | "deadline_at" | "question_count";
-
-function getSortValue(t: Task, field: TaskSortField): number {
-	if (field === "created_at") return new Date(t.createdAt).getTime();
-	if (field === "deadline_at") return new Date(t.deadlineAt).getTime();
-	return t.questionCount;
-}
-
-function sortTasks(tasks: Task[], field: TaskSortField, dir: "asc" | "desc"): Task[] {
-	const mul = dir === "asc" ? 1 : -1;
-	return [...tasks].sort((a, b) => mul * (getSortValue(a, field) - getSortValue(b, field)));
-}
-
-function applyFilters(tasks: Task[], params: TaskFilterParams): Task[] {
-	const q = params.q?.trim().toLowerCase();
-	return tasks.filter((t) => {
-		if (q && !t.name.toLowerCase().includes(q)) return false;
-		if (params.item && t.item.id !== params.item) return false;
-		if (params.company && t.item.companyId !== params.company) return false;
-		return true;
-	});
-}
-
-function applySortIfAny(tasks: Task[], params: TaskFilterParams): Task[] {
-	if (!params.sort) return tasks;
-	return sortTasks(tasks, params.sort, params.dir ?? "asc");
-}
-
-const COLUMN_PAGE_SIZE = 20;
-const LIST_PAGE_SIZE = 20;
-
-// --- Mock API: board ---
-
-export interface BoardColumn {
-	results: Task[];
-	next: string | null;
-	count: number;
-}
-
-export interface TaskBoardResponse {
-	assigned?: BoardColumn;
-	in_progress?: BoardColumn;
-	completed?: BoardColumn;
-	archived?: BoardColumn;
-	results?: Task[];
-	next?: string | null;
-}
-
-export interface FetchTaskBoardParams {
-	q?: string;
-	item?: string;
-	company?: string;
-	sort?: TaskSortField;
-	dir?: "asc" | "desc";
-	column?: TaskStatus;
-	cursor?: string;
-}
-
-function buildColumn(status: TaskStatus, params: FetchTaskBoardParams, cursor?: string): BoardColumn {
-	const filterParams: TaskFilterParams = {
-		q: params.q,
-		item: params.item,
-		company: params.company,
-		sort: params.sort,
-		dir: params.dir,
-	};
-	const filtered = applyFilters(tasksStore, filterParams).filter((t) => t.status === status);
-	const sorted = applySortIfAny(filtered, filterParams);
-	const page = paginate({
-		items: sorted,
-		cursor,
-		limit: COLUMN_PAGE_SIZE,
-		getId: (t) => t.id,
-	});
-	return {
-		results: page.items.map(cloneTask),
-		next: page.nextCursor,
-		count: filtered.length,
-	};
-}
-
-export async function fetchTaskBoardMock(params: FetchTaskBoardParams = {}): Promise<TaskBoardResponse> {
-	await delay();
-	if (params.column) {
-		const col = buildColumn(params.column, params, params.cursor);
-		return { results: col.results, next: col.next };
-	}
-	return {
-		assigned: buildColumn("assigned", params),
-		in_progress: buildColumn("in_progress", params),
-		completed: buildColumn("completed", params),
-		archived: buildColumn("archived", params),
-	};
-}
-
-// --- Mock API: list ---
-
-export interface FetchTasksParams {
-	page?: number;
-	page_size?: number;
-	q?: string;
-	item?: string;
-	company?: string;
-	statuses?: TaskStatus[];
-	sort?: TaskSortField;
-	dir?: "asc" | "desc";
-}
-
-export interface TaskListResponse {
-	count: number;
-	results: Task[];
-	next: string | null;
-	previous: string | null;
-}
-
-export async function fetchAllTasksMock(): Promise<Task[]> {
-	await delay();
-	return tasksStore.map(cloneTask);
-}
-
-export async function fetchTasksMock(params: FetchTasksParams = {}): Promise<TaskListResponse> {
-	await delay();
-	const filterParams: TaskFilterParams = {
-		q: params.q,
-		item: params.item,
-		company: params.company,
-		sort: params.sort,
-		dir: params.dir,
-	};
-	let filtered = applyFilters(tasksStore, filterParams);
-	if (params.statuses && params.statuses.length > 0) {
-		const allowed = new Set(params.statuses);
-		filtered = filtered.filter((t) => allowed.has(t.status));
-	}
-	const sorted = applySortIfAny(filtered, filterParams);
-	const pageSize = params.page_size ?? LIST_PAGE_SIZE;
-	const page = params.page ?? 1;
-	const start = (page - 1) * pageSize;
-	const end = start + pageSize;
-	const slice = sorted.slice(start, end);
-	return {
-		count: filtered.length,
-		results: slice.map(cloneTask),
-		next: end < filtered.length ? `page=${page + 1}` : null,
-		previous: page > 1 ? `page=${page - 1}` : null,
-	};
-}
-
-// --- Mock API: detail ---
-
-export async function fetchTaskMock(id: string): Promise<Task> {
-	await delay();
-	const idx = findTaskIndex(id);
-	if (idx === -1) throw new Error(`Task ${id} not found`);
-	return cloneTask(tasksStore[idx]);
-}
-
-// --- Mock API: status change ---
-
-export async function changeTaskStatusMock(
-	id: string,
-	data: { status: TaskStatus; completedResponse?: string },
-): Promise<Task> {
-	await delay();
-	const idx = requireTaskIdx(id);
-	const current = tasksStore[idx];
-	const updated: Task = {
-		...current,
-		status: data.status,
-		completedResponse: data.completedResponse ?? current.completedResponse,
-		updatedAt: new Date().toISOString(),
-	};
-	if (data.status === "archived" && current.status !== "archived") {
-		updated.statusBeforeArchive = current.status;
-	}
-	tasksStore[idx] = updated;
-	return cloneTask(updated);
-}
-
-// --- Mock API: attachments ---
-
-function fileExtension(name: string): string {
-	const lastDot = name.lastIndexOf(".");
-	return lastDot === -1 ? "" : name.slice(lastDot + 1).toLowerCase();
-}
-
-export async function uploadTaskAttachmentsMock(id: string, files: File[]): Promise<Attachment[]> {
-	await delay();
-	const idx = requireTaskIdx(id);
-	const created: Attachment[] = files.map((file) => ({
-		id: nextId("att"),
-		fileName: file.name,
-		fileSize: file.size,
-		fileType: fileExtension(file.name),
-		contentType: file.type || "application/octet-stream",
-		fileUrl: createBlobUrl(file),
-		uploadedAt: new Date().toISOString(),
-	}));
-	const current = tasksStore[idx];
-	tasksStore[idx] = {
-		...current,
-		attachments: [...current.attachments, ...created],
-		updatedAt: new Date().toISOString(),
-	};
-	return created.map((a) => ({ ...a }));
-}
-
-export async function deleteTaskAttachmentMock(id: string, attachmentId: string): Promise<void> {
-	await delay();
-	const idx = requireTaskIdx(id);
-	const current = tasksStore[idx];
-	tasksStore[idx] = {
-		...current,
-		attachments: current.attachments.filter((a) => a.id !== attachmentId),
-		updatedAt: new Date().toISOString(),
-	};
-}
+export {
+	changeTaskStatusMock,
+	deleteTaskAttachmentMock,
+	uploadTaskAttachmentsMock,
+} from "./tasks-mock/mutations";
+export {
+	fetchAllTasksMock,
+	fetchTaskBoardMock,
+	fetchTaskMock,
+	fetchTasksMock,
+} from "./tasks-mock/queries";
+export { _getAllTasks, _resetTasksStore, _setTasks } from "./tasks-mock/store";
